verify correct splitting behavior of arrays.
The array is split into two parts. If it is of
even length, it is split evenly. If it is of odd
length, it is split with the right half having
one more element than the left. This was verified
by printing it splitting arrays up to a length of
6 elements.